+++ CarShowroom/src/main/java/com/tuantran/CarShowroom/service/SegmentService.java
@@ -9,10 +9,10 @@ import org.springframework.data.jpa.domain.Specification;
 import java.util.List;
 
 public interface SegmentService {
-    Page<SegmentResponse> findByBrand(int brandId, Pageable pageable);
+    Page<SegmentResponse> findByBrand(long brandId, Pageable pageable);
     List<SegmentResponse> findAll();
     Page<SegmentResponse> findAll(Pageable pageable);
-    SegmentResponse findById(int id);
+    SegmentResponse findById(long id);
     Page<SegmentResponse> findAll(Specification<Segment> specification, Pageable pageable);
-    Page<SegmentResponse> findByBrand(int brandId, Specification<Segment> specification, Pageable pageable);
+    Page<SegmentResponse> findByBrand(long brandId, Specification<Segment> specification, Pageable pageable);
 }
